+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kCDsp/Dmc4kCDspAdapterSettings.cs
@@ -1,9 +1,10 @@
-﻿using ICD.Connect.Settings.Attributes;
+﻿using ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kCDspBase;
+using ICD.Connect.Settings.Attributes;
 
 namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kCDsp
 {
 	[KrangSettings("Dmc4kCDsp", typeof(Dmc4kCDspAdapter))]
-	public sealed class Dmc4kCDspAdapterSettings : AbstractInputCardSettings
+	public sealed class Dmc4kCDspAdapterSettings : AbstractDmc4kCDspBaseAdapterSettings
 	{
 	}
 }