+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kHd/Dmc4kHdAdapterSettings.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdBase;
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHd
+{
+	[KrangSettings("Dmc4kHd", typeof(Dmc4kHdAdapter))]
+	public sealed class Dmc4kHdAdapterSettings : AbstractDmc4kHdBaseAdapterSettings
+	{
+	}
+}