+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kHdBase/AbstractDmc4kHdBaseAdapter.cs
@@ -1,0 +1,12 @@
+﻿namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdBase
+{
+#if SIMPLSHARP
+	public abstract class AbstractDmc4kHdBaseAdapter<TCard, TSettings> : AbstractInputCardAdapter<TCard, TSettings>
+		where TCard : Crestron.SimplSharpPro.DM.Cards.Dmc4kHdBase
+#else
+	public abstract class AbstractDmc4kHdBaseAdapter<TSettings> : AbstractInputCardAdapter<TSettings>
+#endif
+		where TSettings : IInputCardSettings, new()
+	{
+	}
+}