+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kHdDsp/Dmc4kHdDspAdapterSettings.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdDspBase;
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdDsp
+{
+	[KrangSettings("Dmc4kHdDsp", typeof(Dmc4kHdDspAdapter))]
+	public sealed class Dmc4kHdDspAdapterSettings : AbstractDmc4kHdDspBaseAdapterSettings
+	{
+	}
+}