+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kzCDsp/Dmc4kzCDspAdapterSettings.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kCDspBase;
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kzCDsp
+{
+	[KrangSettings("Dmc4kzCDsp", typeof(Dmc4kzCDspAdapter))]
+	public sealed class Dmc4kzCDspAdapterSettings : AbstractDmc4kCDspBaseAdapterSettings
+	{
+	}
+}