+++ ICD.Connect.Routing.CrestronPro/Cards/Outputs/Dmc4kCoHdBase/AbstractDmc4kCoHdBaseAdapterSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.CrestronPro.Cards.Outputs.Dmc4kCoHdBase
+{
+	public abstract class AbstractDmc4kCoHdBaseAdapterSettings : AbstractOutputCardSettings
+	{
+	}
+}