+++ ICD.Connect.Routing.Extron/Properties/AssemblyInfo.cs
@@ -3,6 +3,6 @@ using System.Reflection;
 [assembly: AssemblyTitle("ICD.Connect.Routing.Extron")]
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Routing.Extron")]
-[assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
-[assembly: AssemblyVersion("14.1.0.0")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2019")]
+[assembly: AssemblyVersion("14.2.0.0")]