+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [14.2.0] - 2019-02-14
+### Added
+ - Added DM-TX-4KZ-202-C and DM-TX-4KZ-302-C transmitters
+ - Added DMC-4K-HD and DMC-4KZ-HD input cards
+ - Added DMC-4KZ-C-DSP input card
+ - Added DMC-4KZ-C input card
+ - Added DMC-4KZ-HD-DSP input card
+ - Added DMC-4KZ-CO-HD output card
+
 ## [14.1.0] - 2019-01-29
 ### Added
  - DMC-SO-HD DM 8G Fiber Output Card
@@ -20,6 +29,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [14.0.0] - 2019-01-10
 ### Added
  - Added port configuration features to routing devices
+
+## [13.2.0] - 2019-02-13
+### Added
+ - Added DmcCoHd output card variant
 
 ## [13.1.1] - 2019-01-22
 ### Changed